requirements.txt
fastapi==0.115.12
supabase==2.15.2
pydantic==2.11.5
pydantic-settings==2.9.1
google-cloud-secret-manager==2.23.3
python-dotenv==1.1.0
python-jose==3.5.0
uvicorn==0.34.0
email-validator==2.2.0